port/cpl_virtualmem.cpp: MSVC 32 warning fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@31621 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_virtualmem.cpp
@@ -2104,7 +2104,7 @@ CPLVirtualMem *CPLVirtualMemDerivedNew(CPLVirtualMem* pVMemBase,
     ctxt->eAccessMode = pVMemBase->eAccessMode;
     ctxt->pData = (GByte*) pVMemBase->pData + nOffset;
     ctxt->pDataToFree = NULL;
-    ctxt->nSize = nSize;
+    ctxt->nSize = static_cast<size_t>(nSize);
     ctxt->nPageSize = pVMemBase->nPageSize;
     ctxt->bSingleThreadUsage = pVMemBase->bSingleThreadUsage;
     ctxt->pfnFreeUserData = pfnFreeUserData;
